requirements.txt
python-dotenv==1.0.0

pyyaml==6.0.1

langchain==0.1.0

langchain-openai==0.0.2

langchain-anthropic==0.0.1

openai==1.7.0

pandas>=2.2.0

numpy>=1.26.0

faker==22.0.0

scikit-learn>=1.4.0

matplotlib>=3.8.0

seaborn>=0.13.0

plotly>=5.18.0

tqdm>=4.66.0

jupyter>=1.0.0

ipykernel>=6.27.0

pytest>=7.4.0

pytest-cov>=4.1.0

pydantic>=2.5.0

rapidfuzz>=2.14.0
